Fix #57 and a bug in URL decoding
The patch contains several fixes to the following issues:

(1) #57 NPE Again https://github.com/walmartlabs/mupd8/issues/57
  Case (a): GET / -> ArrayOutOfBoundsException
  Solution: The array boundary is first checked before calling tok(2). For example,
  if (tok.length <= 2 || tok(1) == "favicon.ico") None

  Case (b): GET /mupd8/slate/app/<unrecognized Updater>/<key> -> NoSuchElementException
  Solution: use a try-catch block to catch the exception return None, so
that the client will see a 404 NOT FOUND response.

  Case (c): GET //mupd8/… and GET /mupd8/… yield different results
  Solution: the redundant path separators should be collapsed so
that //mupd8/ and /mupd8/ and //mupd8///// will look same to the server
and will return the same response. normalizeURL() in the patch will do this job.

(2) the slate key in URL is not correctly decoded when it contains path
separator '/'
  Solution: since the slate key is typically percent-coded in the URL,
it would be misleading to upfrontly decode the entire URL then parse out
the slate key. The better way to do this is to reverse that order: (1)
parse/split the entire URL by '/'. (2) decode the slate key.

Testing: To (1), random URLs (both valid and invalid) are used to
test via Curl. To (2), the patch runs through the test cases provided by the
issue reporter. In addition, TestApp is also run against this patch.